+++ cmake_format/__init__.py
@@ -3,4 +3,4 @@ Parse cmake listfiles and format them nicely
 """
 from __future__ import unicode_literals
 
-VERSION = '0.6.4'
+VERSION = '0.6.5'